// frontend/src/shared/lib/api/model/nameClass.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * prot-cellar
 * OpenAPI spec version: 0.1.0
 */

export type NameClass = typeof NameClass[keyof typeof NameClass];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NameClass = {
  scientific_name: 'scientific_name',
  common_name: 'common_name',
  genbank_common_name: 'genbank_common_name',
  synonym: 'synonym',
  authority: 'authority',
  equivalent_name: 'equivalent_name',
  acronym: 'acronym',
  blast_name: 'blast_name',
  uniprot_mnemonic: 'uniprot_mnemonic',
} as const;