// frontend/src/shared/lib/api/model/organismSource.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * prot-cellar
 * OpenAPI spec version: 0.1.0
 */

export type OrganismSource = typeof OrganismSource[keyof typeof OrganismSource];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrganismSource = {
  ncbi: 'ncbi',
  gtdb: 'gtdb',
  local: 'local',
} as const;