// frontend/src/shared/lib/api/model/proteomeType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * prot-cellar
 * OpenAPI spec version: 0.1.0
 */

export type ProteomeType = typeof ProteomeType[keyof typeof ProteomeType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProteomeType = {
  reference: 'reference',
  representative: 'representative',
  redundant: 'redundant',
  excluded: 'excluded',
} as const;